chore(driver-adapters): rename "flavour" to "provider"
