+++ packages/adapter-neon/src/neon.ts
@@ -24,7 +24,7 @@ type PerformIOResult = neon.QueryResult<any> | neon.FullQueryResults<ARRAY_MODE_
  * Base class for http client, ws client and ws transaction
  */
 abstract class NeonQueryable implements Queryable {
-  readonly flavour = 'postgres'
+  readonly provider = 'postgres'
 
   async queryRaw(query: Query): Promise<Result<ResultSet>> {
     const tag = '[js::query_raw]'
