+++ packages/client/tests/functional/driver-adapters/_utils/mock-adapter.ts
@@ -1,7 +1,7 @@
 import { DriverAdapter } from '@prisma/driver-adapter-utils'
 
 import { Providers as Provider } from '../../_utils/providers'
-import { flavourFromProvider } from './flavour'
+import { getDriverAdaptersProvider } from './provider'
 
 export const mockAdapterErrors = {
   queryRaw: new Error('Not implemented: queryRaw'),
@@ -15,7 +15,7 @@ export const mockAdapterErrors = {
  */
 export function mockAdapter(provider: Provider): DriverAdapter {
   return {
-    flavour: flavourFromProvider(provider),
+    provider: getDriverAdaptersProvider(provider),
     queryRaw: () => Promise.reject(mockAdapterErrors.queryRaw),
     executeRaw: () => Promise.reject(mockAdapterErrors.executeRaw),
     startTransaction: () => Promise.reject(mockAdapterErrors.startTransaction),
