+++ packages/client/tests/functional/driver-adapters/adapter-nullability/_matrix.ts
@@ -1,4 +1,4 @@
 import { defineMatrix } from '../../_utils/defineMatrix'
-import { driverAdaptersTestProviders } from '../_utils/flavour'
+import { driverAdaptersTestProviders } from '../_utils/provider'
 
 export default defineMatrix(() => [driverAdaptersTestProviders])
