+++ packages/client/tests/functional/driver-adapters/enforce-preview-feature/_matrix.ts
@@ -1,5 +1,5 @@
 import { defineMatrix } from '../../_utils/defineMatrix'
-import { driverAdaptersTestProviders } from '../_utils/flavour'
+import { driverAdaptersTestProviders } from '../_utils/provider'
 
 export default defineMatrix(() => [
   driverAdaptersTestProviders,
