+++ packages/driver-adapter-utils/src/binder.ts
@@ -28,7 +28,7 @@ export const bindAdapter = (adapter: DriverAdapter): ErrorCapturingDriverAdapter
     errorRegistry,
     queryRaw: wrapAsync(errorRegistry, adapter.queryRaw.bind(adapter)),
     executeRaw: wrapAsync(errorRegistry, adapter.executeRaw.bind(adapter)),
-    flavour: adapter.flavour,
+    provider: adapter.provider,
     startTransaction: async (...args) => {
       const result = await startTransaction(...args)
       return result.map((tx) => bindTransaction(errorRegistry, tx))
@@ -41,7 +41,7 @@ export const bindAdapter = (adapter: DriverAdapter): ErrorCapturingDriverAdapter
 // execution is delegated to napi.rs.
 const bindTransaction = (errorRegistry: ErrorRegistryInternal, transaction: Transaction): Transaction => {
   return {
-    flavour: transaction.flavour,
+    provider: transaction.provider,
     options: transaction.options,
     queryRaw: wrapAsync(errorRegistry, transaction.queryRaw.bind(transaction)),
     executeRaw: wrapAsync(errorRegistry, transaction.executeRaw.bind(transaction)),
